requirements.txt
Django==4.1.7
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.2
rest_framework_simplejwt==0.0.2
pillow==10.0.0
django-redis==5.3.0
redis==4.6.0
channels==4.0.0
pytest_django==4.5.2
